+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## [Unreleased]
+## [2.5] - 2022-01-04
 ### Added
 - `Slf4jProxy`, `LoggerException`
 ### Changed
 - `Slf4JLogger`, `Slf4JLoggerException`, `DefaultLoggerFactory`: improved reflection handling
 - `ConsoleLogger`: Added error handling
+- `JavaLogger`: Added use of `LogListener`
 
 ## [2.4] - 2021-09-29
 ### Changed